feat(TDD cycle: Red stage): Add failing test to check if day exists.
This new set of tests describe  a desired behavior for the
`DaysOfWeekManager` class: "Both should return the correct answer
depending on wether the day exists in the list or not (	`true` or
`false` respectively).

This test "should fail!", because the code to make it pass hasn't been
written yet (`dayExists` method doesn't exist yet).