feat(TDD cycle: Red stage): Add failing test to empty list.
This new test describe a desired behavior for the `DaysOfWeekManager`
class: "It should succesfully clears the list of days".

This test "should fail!", because the code to make it pass hasn't been
written yet (`emptyList` method doesn't exist yet).

Hamcrest Library is used to assert that the list is empty.